Treat X11 handles as size_t
Treat the XID, VisualID, Window, and RROutput types as size_t, as
they seem to be defined as unsigned long and are 32-bit for x86
and 64-bit for x86_64.

+++ framework/format/platform_types.h
@@ -710,9 +710,9 @@ static VKAPI_ATTR VkBool32 VKAPI_CALL vkGetPhysicalDeviceXcbPresentationSupportK
 
 typedef VkFlags VkXlibSurfaceCreateFlagsKHR;
 
-typedef uint32_t XID;
-typedef uint32_t VisualID;
-typedef XID      Window;
+typedef size_t XID;
+typedef size_t VisualID;
+typedef XID    Window;
 
 struct Display;
